# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/archives/{id}/download": {
            "get": {
                "description": "Скачивает архив, связанный с задачей, по её ID",
                "produces": [
                    "application/zip"
                ],
                "tags": [
                    "archives"
                ],
                "summary": "Скачать архив по ID задачи",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Неверный ID задачи",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Архив не найден",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка открытия файла",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/tasks/create": {
            "post": {
                "description": "Создает новую задачу и возвращает её ID",
                "tags": [
                    "tasks"
                ],
                "summary": "Создать новую задачу",
                "responses": {
                    "201": {
                        "description": "Задача успешно создана с ID: {id}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка при создании задачи",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/tasks/{id}/add-link": {
            "post": {
                "description": "Добавляет ссылку к задаче по её ID",
                "tags": [
                    "tasks"
                ],
                "summary": "Добавить ссылку к задаче",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ссылка для добавления",
                        "name": "link",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ссылка успешно добавлена к задаче",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный ID задачи или пустая ссылка",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка при добавлении ссылки к задаче",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/tasks/{id}/status": {
            "get": {
                "description": "Возвращает статусы задачи по её ID. В случае, когда ни один файл не удалось скачать, архив не будет возвращён.\nЕсли задача завершена успешно/удалось установить хоть один файл на момент завершения, возвращает ссылку на скачивание архива",
                "tags": [
                    "tasks"
                ],
                "summary": "Получить статусы задачи",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статусы задачи успешно получены",
                        "schema": {
                            "$ref": "#/definitions/internal_routes.GetStatusesResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID задачи",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка при получении или сериализации статусов задачи",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_routes.GetStatusesResponse": {
            "type": "object",
            "properties": {
                "download_link": {
                    "type": "string"
                },
                "task": {
                    "$ref": "#/definitions/internal_routes.Task"
                }
            }
        },
        "internal_routes.Task": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}